#2 cleaning and refactoring
