+++ src/main/kotlin/com/github/christophpickl/payconiq/misc.kt
